Updated dataset and changes in the do-file
the following changes were made: 1) the blinded TrialArm variable is in the updated dataset, as defined by the datamanager, 2) the variables "us" and "lab" have now been defined by the data manager and already exist in the updated dataset, so 3) the lines in the do-file that were to define these 2 variables have now been removed. 4) the data signature was revised to correspond to the updated dataset.